bitgenius-back/bitgenius-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

firebase-admin

requests
httpx

google-generativeai

python-multipart
python-jose
